Update ApplicationOAuthProvider.cs
Fixes issue within OWIN when using multiple allowed origins for cross-origin resource sharing. This affected calls to the WebAPI "/token" endpoint when requesting a bearer token. Thus, what would occur before was that the key "Access-Control-Allow-Origin" would return a value that was set to the entire string, including comma that was used within the config file. Which would yield an error when trying to access from any of the origins in the config file. Everything got treated as one long URI instead of two independent URIs.

The fix gets the origin from the config, but then now also grabs the origin value from the request before checking to make sure the allowed origins string contains the origin from the request. Then dynamically serves back that origin as an allowed origin within the response headers.

Comments are added to keep things neat with the original code still in place with an additional line added to demonstrate handling all origins with a "*" value passed with the "Access-Control-Allow-Origin" key.